requirements.txt
python-dotenv==1.0.1python-dotenv==1.0.1

unstructured-client>=0.42.0unstructured-client>=0.42.0

python-docx>=1.1.2python-docx>=1.1.2

requests>=2.25.0requests>=2.25.0

chromadb>=0.5.5chromadb>=0.5.5

sentence-transformers>=2.6.0sentence-transformers>=2.6.0

groq>=0.4.0groq>=0.4.0

langchain-text-splitters==0.0.1langchain-text-splitters==0.0.1

websockets>=12.0websockets>=12.0

tf-keras>=2.15.0